requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
